Move APIKeys model from imports/{server -> lib}
This prepares us to expose information about APIKeys to the client, so
we can provide UI for API key management.
